Create a basic scene
We can only have 1 instance of World, or Scene at a time. Therefore, Scene should be remained to something like Game, or such, because it cannot be instantiated twice. We should then extend GreenfootImage to create a GameScreen or such, which will represent the objects and text that we will use.